net: ipc_router: Fix memory leaks when releasing a remote port
In ipc_router_set_conn(), memory is allocated to a local port for the
conn_info linked list, but the memory for the local ports is never
freed when the remote port is released.

Delete the local ports and free all of their memory when releasing a remote
port to fix the memory leaks. This is just like how memory is freed for the
remote_tx ports.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
(cherry picked from commit df142e91cdddb22dec9ec83a72acecc12fbad100)
Signed-off-by: Jimbo77 <chupyalcohol@gmail.com>